Create complex shapes cast with surface modeling
Ornament - Thesis includes a description of the procedure necessary to create a 3D solid model based on the actual artistic cast. The 3D solid contained complex shapes surface elements and was created in CAD 3D software SolidWorks. SolidWorks is in an educational version, edition 2016-2017. The presented material is based on the knowledge contained in the software help, so-called tutorial and classes held in the course of engineering studies and also the book [1].
In the theoretical part it is described computer programs CAD 3D, in which it is possible to use surface objects. The theoretical part contains a description of modeling techniques parts created in SolidWorks. The last part of this section describes methods for creating 3D models by: creating solids, 3D scanning.
In the practical part of the course of action was concluded, consisting of three stages: the first stage - the methodology of measuring physical elements, the second stage - the methodology of creating elements and the third stage in the form of photographs with descriptions that allows accurate tracing of all necessary operations to create solids.